// clickify-generated/src/router/typings/Pages.ts
export enum Pages {
  ERROR_PAGE = "errorPage",
  LANDING_PAGE = "landingPage",
  LOGIN_PAGE = "loginPage",
  SIGNUP_PAGE = "signupPage",
  PRODUCTS_PAGE = "products", // /products
  PRODUCT_DETAILS_PAGE = "productDetails", // /products/:id
  CART_PAGE = "cart", // /cart
  RECAP_PAGE = "recap", // /recap
  ORDER_HISTORY_PAGE = "orderHistory", // /order-history
  ACCOUNT_PAGE = "account", // /account
  ACCOUNT_ADDRESSES_PAGE = "accountAddresses", // /account/addresses
  ACCOUNT_PAYEMENTS_PAGE = "accountPayements", // /account/payements
  ACCOUNT_STORE_PAGE = "accountStore", // /account/store
  ABOUT_US_PAGE = "aboutUs", // /about-us
  SUPPORT_PAGE = "support", // /support
}